refactor: move HTTP API contracts out of core/interfaces/Api.ts
Moved APIGet, APIPost, GetRoute, PostRoute, InvitationResult, and
InvitationSummaryRow from src/core/interfaces/Api.ts to a new
src/core/api/ApiContracts.ts. The original file now contains only
pure domain utilities (Params, LanguageTimestamp,
encodeLanguageTimestamps, decodeLanguageTimestamps).

Updated all consumers:
- src/core/api/WebAPIClient.ts
- src/server/api/WebAPI.ts
- src/desktop/DesktopAPIServer.ts
- src/desktop/WebAPIClientForDesktop.ts
- src/frontend/common/api/RequestContext.ts
- src/frontend/desktopFrontend/desktopAPIClient.ts
- src/frontend/web/auth/redeemInvitationThunks.ts
- src/frontend/web/invitations/invitationThunks.ts
- src/frontend/web/invitations/invitationsListThunks.ts

Added apiContracts.test.ts to verify the new location and assert
that core/interfaces/Api no longer exports the HTTP-specific types.
